\name{GenBin}
\alias{GenBin}
\docType{data}
\title{
Simulated genetic data with binary phenotypes
}
\description{
Simulated genetic data at 15 SNPs, together with simulated binary phenotypes
}
\usage{data(GenBin)}
\format{
	Numeric matrix with column names. First column is the simulated phenotype, columns 2-16 are the simulated genotypes.
}
\source{
	Genotypes were simulated using FREGENE.
}
\references{
Fregene: Simulation of realistic sequence-level data in populations and ascertained samples Chadeau-Hyam, M. \emph{et al}, 
2008, BMC Bioinformatics 9:364
}
\examples{
data(GenBin)
}
\keyword{datasets}
