\name{GenCont}
\alias{GenCont}
\docType{data}
\title{
Simulated genetic data with continuous outcomes
}
\description{
Simulated genetic data at 12 SNPs, together with simulated continuous outcomes.
}
\usage{data(GenCont)}
\format{
	Numeric matrix with column names. First column is the simulated phenotype, columns 2-13 are the simulated genotypes.
}
\details{
	Genotypes were simulated using FREGENE.
}
\references{
Fregene: Simulation of realistic sequence-level data in populations and ascertained samples Chadeau-Hyam, M. \emph{et al}, 
2008, BMC Bioinformatics 9:364
}
\examples{
data(GenCont)
}
\keyword{datasets}
