% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{lm.ridge}
\alias{lm.ridge}
\title{Square}
\usage{
lm.ridge(formula, data, subset, na.action, lambda = 0, model = FALSE,
  x = FALSE, y = FALSE, contrasts = NULL, ...)
}
\arguments{
\item{x}{input value}
}
\value{
the square of x
}
\description{
Gets the square of a number
}
\examples{
square(3)

}
