% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ridge_model.R
\name{print.ridge_model}
\alias{print.ridge_model}
\title{Print a ridge regression model}
\usage{
\method{print}{ridge_model}(x, ...)
}
\arguments{
\item{x}{A fitted ridge regression model.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns the model.
}
\description{
Print a ridge regression model
}
