\name{ridge-package}
\alias{ridge-package}
\alias{ridge}
\docType{package}
\title{
ridge package
}
\description{
R package for fitting linear and logistic ridge regression models.
}
\details{
This package contains functions for fitting linear and logistic ridge regression models. 

For a complete
list of functions, use \code{library(help="ridge")}.

}
\author{
Erika Cule
}
\keyword{ package }
