% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\docType{package}
\name{ridge}
\alias{ridge}
\title{ridge: Linear and logistic ridge regression functions.}
\description{
Additionally includes special functions for 
genome-wide single-nucleotide polymorphism (SNP) data.
}
