% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_fit.R
\name{ridge_fit}
\alias{ridge_fit}
\title{Fit a ridge regression model}
\usage{
ridge_fit(x, y, lambda = 1)
}
\arguments{
\item{x}{A numeric matrix of predictors.}

\item{y}{A numeric vector of responses.}

\item{lambda}{A non-negative regularization parameter.}
}
\value{
A list containing the fitted coefficients and the value of
the regularization parameter.
}
\description{
Fits a ridge regression model using a regularization parameter.
}
\examples{
x <- matrix(rnorm(20), ncol = 2)
y <- rnorm(10)
model <- ridge_fit(x, y, lambda = 1)
model

}
