% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ridgereg.r
\docType{class}
\name{rigdereg-class}
\alias{ridgereg}
\alias{rigdereg-class}
\title{Function for ridge regression}
\arguments{
\item{formula}{formula.}

\item{data}{data.frame.}
}
\value{
A linreg object.
}
\description{
Function for ridge regression
}
\section{Fields}{

\describe{
\item{\code{formula}}{A formula}

\item{\code{data}}{A data.frame}

\item{\code{bhat}}{A matrix}

\item{\code{yhat}}{A matrix}
}}
\examples{
ridgereg1 <- ridgereg(formula = Petal.Length ~ Species, data = iris)
ridgereg1$print()
ridgereg1$coeff()
ridgereg1$pred()
## We decide to use QR decompositon.
}

