% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{solve_ridge}
\alias{solve_ridge}
\title{Solve Ridge Regression}
\usage{
solve_ridge(x, y, lambda)
}
\arguments{
\item{x}{a matrix of predictors}

\item{y}{a vector of responses}

\item{lambda}{a scalar representing the penalty parameter}
}
\description{
This function solves ridge regression.
}
\details{
Note that this function assumes that the X is standardized.
}
