% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge2.R
\name{solve_ridge_2}
\alias{solve_ridge_2}
\title{Solve Ridge Regression 2}
\usage{
solve_ridge_2(x, y, lambda)
}
\arguments{
\item{x}{a matrix of predictors}

\item{y}{a vector of responses}

\item{lambda}{a scalar representing the penalty parameter}
}
\description{
This function solves ridge regression using glmnet.
}
