% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{creates a plot which visualize a large dataset}
\usage{
visualize_airport_delays()
}
\value{
returns plot
}
\description{
creates a plot which visualize a large dataset
}
\examples{
visualize_airport_delays()

}
