\name{Hald}
\alias{Hald}
\docType{data}
\title{
Hald data
}
\description{
The Hald data as used by Hoerl, Kennard and Baldwin (1975). 
These data are also in package \code{wle}.
}
\usage{data(Hald)}
\format{
	Numeric matrix.
}
\details{
The first column is the response and the remaining four columns are the predictors.
}
\references{
Ridge Regression: some simulations, Hoerl, A. E. \emph{et al}, 1975, Comm Stat Theor Method 4:105
}
\examples{
data(Hald)
}
\keyword{datasets}
