\docType{package}
\name{ridge-package}
\alias{ridge-package}
\title{ridge-package description}
\description{
R package for fitting linear and logistic ridge regression models.
}
\details{
This package contains functions for fitting linear and logistic ridge
regression models, including functions for fitting linear and logistic
ridge regression models for genome-wide SNP data supplied as file names
when the data are too big to read into R.

For a complete
list of functions, use \code{help(package="ridge")}.
}
\author{
Steffen Moritz, Erika Cule
}
